requirements.txt
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
folium>=0.17.0
PyYAML>=6.0.1
uvicorn>=0.29.0
pillow-heif>=1.3.0
